requirements.txt
Django==4.1.7
django-filter==22.1
djangorestframework==3.14.0
ipython==8.10.0
Jinja2==3.1.2
simplejson==3.18.3
six==1.16.0
uritemplate==4.1.1
urllib3==1.26.14
wcwidth==0.2.6
